[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14376 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -10,8 +10,8 @@ public class Main {
 
     public static void main(String[] args) throws IOException {
         StringTokenizer tk = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(tk.nextToken());
-        int b = Integer.parseInt(tk.nextToken());
+        long a = Long.parseLong(tk.nextToken());
+        long b = Long.parseLong(tk.nextToken());
         System.out.println((a+b) * (a-b));
     }
 }
